+++ c++/mfdfa.h
@@ -9,14 +9,14 @@
 class MFDFA
 {
 public:
-    explicit MFDFA(double *h_y, int yLen);
+    explicit MFDFA(float *h_y, int yLen);
     ~MFDFA();
-    void computeFlucVec(int *winSizes, int nWins, double *qVals, int nq, double *hq, int threads, bool revSeg=false);
-    void computeMultifractalSpectrum(int *winSizes, int nWins, double *qVals, int nq, double *a, double *fa, int threads, bool revSeg=false);
+    void computeFlucVec(int *winSizes, int nWins, float *qVals, int nq, float *hq, int threads, bool revSeg=false);
+    void computeMultifractalSpectrum(int *winSizes, int nWins, float *qVals, int nq, float *a, float *fa, int threads, bool revSeg=false);
 private:
     cudaError_t cudaErr;
-    double *d_y;
-    double *d_t;
+    float *d_y;
+    float *d_t;
     int len = 0;
 };
